+++ packages/dht/src/transport/ITransport.ts
@@ -33,4 +33,5 @@ export interface ITransport {
     send(msg: Message, opts?: SendOptions): Promise<void>
     getLocalPeerDescriptor(): PeerDescriptor
     stop(): void | Promise<void>
+    getDiagnosticInfo(): Record<string, unknown>
 }