+++ packages/dht/test/integration/WebrtcConnectorRpc.test.ts
@@ -64,12 +64,12 @@ describe('WebRTC rpc messages', () => {
         rpcCommunicator2.registerRpcNotification(IceCandidate, 'iceCandidate', serverFunctions.iceCandidate)
         rpcCommunicator2.registerRpcNotification(WebrtcConnectionRequest, 'requestConnection', serverFunctions.requestConnection)
 
-        rpcCommunicator1.on('outgoingMessage', (message: RpcMessage) => {
-            rpcCommunicator2.handleIncomingMessage(message)
+        rpcCommunicator1.setOutgoingMessageListener(async (message: RpcMessage) => {
+            rpcCommunicator2.handleIncomingMessage(message, new DhtCallContext())
         })
 
-        rpcCommunicator2.on('outgoingMessage', (message: RpcMessage) => {
-            rpcCommunicator1.handleIncomingMessage(message)
+        rpcCommunicator2.setOutgoingMessageListener(async (message: RpcMessage) => {
+            rpcCommunicator1.handleIncomingMessage(message, new DhtCallContext())
         })
 
         client = toProtoRpcClient(new WebrtcConnectorRpcClient(rpcCommunicator1.getRpcClientTransport()))