+++ packages/dht/test/integration/WebsocketClientConnectorRpc.test.ts
@@ -32,12 +32,12 @@ describe('WebsocketClientConnectorRpc', () => {
             mockWebsocketClientConnectorRpc.requestConnection
         )
 
-        rpcCommunicator1.on('outgoingMessage', (message: RpcMessage) => {
-            rpcCommunicator2.handleIncomingMessage(message)
+        rpcCommunicator1.setOutgoingMessageListener(async (message: RpcMessage) => {
+            rpcCommunicator2.handleIncomingMessage(message, new DhtCallContext())
         })
 
-        rpcCommunicator2.on('outgoingMessage', (message: RpcMessage) => {
-            rpcCommunicator1.handleIncomingMessage(message)
+        rpcCommunicator2.setOutgoingMessageListener(async (message: RpcMessage) => {
+            rpcCommunicator1.handleIncomingMessage(message, new DhtCallContext())
         })
 
         client1 = toProtoRpcClient(new WebsocketClientConnectorRpcClient(rpcCommunicator1.getRpcClientTransport()))