+++ packages/proto-rpc/test/integration/toProtoRpcClient.test.ts
@@ -54,11 +54,11 @@ describe('toProtoRpcClient', () => {
         const helloClient = toProtoRpcClient(new HelloRpcServiceClient(communicator2.getRpcClientTransport()))
 
         // Simulate a network connection, in real life the message blobs would be transferred over a network
-        communicator1.on('outgoingMessage', (msg: RpcMessage, _requestId: string, _callContext?: ProtoCallContext) => {
-            communicator2.handleIncomingMessage(msg)
+        communicator1.setOutgoingMessageListener(async (msg: RpcMessage, _requestId: string, _callContext?: ProtoCallContext) => {
+            communicator2.handleIncomingMessage(msg, new ProtoCallContext())
         })
-        communicator2.on('outgoingMessage', (msg: RpcMessage, _requestId: string, _callContext?: ProtoCallContext) => {
-            communicator1.handleIncomingMessage(msg)
+        communicator2.setOutgoingMessageListener(async (msg: RpcMessage, _requestId: string, _callContext?: ProtoCallContext) => {
+            communicator1.handleIncomingMessage(msg, new ProtoCallContext())
         })
 
         const { greeting } = await helloClient.sayHello({ myName: 'Alice' })
@@ -79,11 +79,11 @@ describe('toProtoRpcClient', () => {
         const wakeUpClient = toProtoRpcClient(new WakeUpRpcServiceClient(communicator2.getRpcClientTransport()))
 
         // Simulate a network connection, in real life the message blobs would be transferred over a network
-        communicator1.on('outgoingMessage', (msg: RpcMessage, _requestId: string, _callContext?: ProtoCallContext) => {
-            communicator2.handleIncomingMessage(msg)
+        communicator1.setOutgoingMessageListener(async (msg: RpcMessage, _requestId: string, _callContext?: ProtoCallContext) => {
+            communicator2.handleIncomingMessage(msg, new ProtoCallContext())
         })
-        communicator2.on('outgoingMessage', (msg: RpcMessage, _requestId: string, _callContext?: ProtoCallContext) => {
-            communicator1.handleIncomingMessage(msg)
+        communicator2.setOutgoingMessageListener(async (msg: RpcMessage, _requestId: string, _callContext?: ProtoCallContext) => {
+            communicator1.handleIncomingMessage(msg, new ProtoCallContext())
         })
 
         wakeUpService.on('wakeUpCalled', async (reason) => {
@@ -107,11 +107,11 @@ describe('toProtoRpcClient', () => {
         const optionalClient = toProtoRpcClient(new OptionalServiceClient(communicator2.getRpcClientTransport()))
 
         // Simulate a network connection, in real life the message blobs would be transferred over a network
-        communicator1.on('outgoingMessage', (msg: RpcMessage, _requestId: string, _callContext?: ProtoCallContext) => {
-            communicator2.handleIncomingMessage(msg)
+        communicator1.setOutgoingMessageListener(async (msg: RpcMessage, _requestId: string, _callContext?: ProtoCallContext) => {
+            communicator2.handleIncomingMessage(msg, new ProtoCallContext())
         })
-        communicator2.on('outgoingMessage', (msg: RpcMessage, _requestId: string, _callContext?: ProtoCallContext) => {
-            communicator1.handleIncomingMessage(msg)
+        communicator2.setOutgoingMessageListener(async (msg: RpcMessage, _requestId: string, _callContext?: ProtoCallContext) => {
+            communicator1.handleIncomingMessage(msg, new ProtoCallContext())
         })
 
         const { someOptionalField } = await optionalClient.getOptional({ someOptionalField: 'something' })
@@ -149,8 +149,8 @@ describe('toProtoRpcClient', () => {
 
         // Simulate a network connection, in real life the message blobs would be transferred over a network
 
-        communicator2.on('outgoingMessage', (msg: RpcMessage, _requestId: string, _callContext?: ProtoCallContext) => {
-            communicator1.handleIncomingMessage(msg)
+        communicator2.setOutgoingMessageListener(async (msg: RpcMessage, _requestId: string, _callContext?: ProtoCallContext) => {
+            communicator1.handleIncomingMessage(msg, new ProtoCallContext())
         })
 
         await wakeUpClient.wakeUp({ reason: 'School' })
@@ -170,7 +170,7 @@ describe('toProtoRpcClient', () => {
         // Simulate a network connection, in real life the message blobs would be transferred over a network
 
         communicator2.setOutgoingMessageListener(async (msg: RpcMessage, _requestId: string, _callContext?: ProtoCallContext) => {
-            communicator1.handleIncomingMessage(msg)
+            communicator1.handleIncomingMessage(msg, new ProtoCallContext())
         })
 
         await wakeUpClient.wakeUp({ reason: 'School' })
@@ -201,11 +201,11 @@ describe('toProtoRpcClient', () => {
         const helloClient = new HelloRpcServiceClient(communicator2.getRpcClientTransport())
 
         // Simulate a network connection, in real life the message blobs would be transferred over a network
-        communicator1.on('outgoingMessage', (msg: RpcMessage, _requestId: string, _callContext?: ProtoCallContext) => {
-            communicator2.handleIncomingMessage(msg)
+        communicator1.setOutgoingMessageListener(async (msg: RpcMessage, _requestId: string, _callContext?: ProtoCallContext) => {
+            communicator2.handleIncomingMessage(msg, new ProtoCallContext())
         })
-        communicator2.on('outgoingMessage', (msg: RpcMessage, _requestId: string, _callContext?: ProtoCallContext) => {
-            communicator1.handleIncomingMessage(msg)
+        communicator2.setOutgoingMessageListener(async (msg: RpcMessage, _requestId: string, _callContext?: ProtoCallContext) => {
+            communicator1.handleIncomingMessage(msg, new ProtoCallContext())
         })
 
         try {
@@ -230,11 +230,11 @@ describe('toProtoRpcClient', () => {
         const wakeUpClient = new WakeUpRpcServiceClient(communicator2.getRpcClientTransport())
 
         // Simulate a network connection, in real life the message blobs would be transferred over a network
-        communicator1.on('outgoingMessage', (msg: RpcMessage, _requestId: string, _callContext?: ProtoCallContext) => {
-            communicator2.handleIncomingMessage(msg)
+        communicator1.setOutgoingMessageListener(async (msg: RpcMessage, _requestId: string, _callContext?: ProtoCallContext) => {
+            communicator2.handleIncomingMessage(msg, new ProtoCallContext())
         })
-        communicator2.on('outgoingMessage', (msg: RpcMessage, _requestId: string, _callContext?: ProtoCallContext) => {
-            communicator1.handleIncomingMessage(msg)
+        communicator2.setOutgoingMessageListener(async (msg: RpcMessage, _requestId: string, _callContext?: ProtoCallContext) => {
+            communicator1.handleIncomingMessage(msg, new ProtoCallContext())
         })
 
         wakeUpService.on('wakeUpCalled', async (_reason) => {