+++ packages/sdk/src/exports.ts
@@ -90,6 +90,7 @@ export {
 } from './contracts/Operator'
 import {
     delegate,
+    undelegate,
     deploySponsorshipContract,
     setupOperatorContract,
     SetupOperatorContractOpts,
@@ -98,9 +99,13 @@ import {
     DeployOperatorContractOpts,
     sponsor,
     stake,
+    unstake,
     getProvider,
     generateWalletWithGasAndTokens,
     DeploySponsorshipContractOpts,
+    getTestTokenContract,
+    getTestAdminWallet,
+    getOperatorContract
 } from './contracts/operatorContractUtils'
 /**
  * @deprecated
@@ -109,13 +114,18 @@ import {
 // eslint-disable-next-line no-underscore-dangle
 const _operatorContractUtils = {
     delegate,
+    undelegate,
     deploySponsorshipContract,
     setupOperatorContract,
     sponsor,
     stake,
+    unstake,
     getProvider,
     generateWalletWithGasAndTokens,
-    deployOperatorContract
+    deployOperatorContract,
+    getTestTokenContract,
+    getTestAdminWallet,
+    getOperatorContract
 }
 // eslint-disable-next-line no-underscore-dangle
 export { _operatorContractUtils }