+++ packages/sdk/test/unit/LocalGroupKeyStore.test.ts
@@ -1,34 +1,23 @@
 import 'reflect-metadata'
 
-import { randomEthereumAddress } from '@streamr/test-utils'
-import { EthereumAddress, toStreamID } from '@streamr/utils'
-import crypto from 'crypto'
+import { randomEthereumAddress, randomUserId } from '@streamr/test-utils'
+import { toStreamID, UserID } from '@streamr/utils'
 import range from 'lodash/range'
-import { DestroySignal } from '../../src/DestroySignal'
-import { PersistenceManager } from '../../src/PersistenceManager'
 import { GroupKey } from '../../src/encryption/GroupKey'
 import { LocalGroupKeyStore } from '../../src/encryption/LocalGroupKeyStore'
-import { getLocalGroupKeyStore, mockLoggerFactory } from '../test-utils/utils'
+import { getLocalGroupKeyStore } from '../test-utils/utils'
 
 describe('LocalGroupKeyStore', () => {
     
-    let clientId: EthereumAddress
-    let publisherId: EthereumAddress
+    let ownerId: UserID
+    let publisherId: UserID
     let store: LocalGroupKeyStore
     let store2: LocalGroupKeyStore
-    let persistenceManager: PersistenceManager
 
     beforeEach(() => {
-        clientId = randomEthereumAddress()
-        publisherId = randomEthereumAddress()
-        store = getLocalGroupKeyStore(clientId)
-        persistenceManager = new PersistenceManager(
-            {
-                getAddress: async () => clientId
-            } as any, 
-            new DestroySignal(),
-            mockLoggerFactory()
-        )
+        ownerId = randomUserId()
+        publisherId = randomUserId()
+        store = getLocalGroupKeyStore(ownerId)
     })
 
     afterEach(async () => {
@@ -36,7 +25,6 @@ describe('LocalGroupKeyStore', () => {
         store = undefined // eslint-disable-line require-atomic-updates
         // @ts-expect-error doesn't want us to unassign, but it's ok
         store2 = undefined // eslint-disable-line require-atomic-updates
-        // TODO trigger destroySignal in persistenceManager
     })
 
     it('can get and set', async () => {
@@ -51,12 +39,12 @@ describe('LocalGroupKeyStore', () => {
         const groupKey = GroupKey.generate()
         await store.set(groupKey.id, publisherId, groupKey.data)
         expect(await store.get(groupKey.id, publisherId)).toEqual(groupKey)
-        expect(await store.get(groupKey.id, randomEthereumAddress())).toBeUndefined()
+        expect(await store.get(groupKey.id, randomUserId())).toBeUndefined()
     })
 
-    it('key stores are clientId specific', async () => {
-        const clientId2 = randomEthereumAddress()
-        store2 = getLocalGroupKeyStore(clientId2)
+    it('key stores are ownerId specific', async () => {
+        const ownerId2 = randomUserId()
+        store2 = getLocalGroupKeyStore(ownerId2)
 
         const groupKey = GroupKey.generate()
         await store.set(groupKey.id, publisherId, groupKey.data)
@@ -68,13 +56,13 @@ describe('LocalGroupKeyStore', () => {
         const groupKey = GroupKey.generate()
         await store.set(groupKey.id, publisherId, groupKey.data)
 
-        const store2 = getLocalGroupKeyStore(clientId)
+        const store2 = getLocalGroupKeyStore(ownerId)
         expect(await store2.get(groupKey.id, publisherId)).toEqual(groupKey)
     })
 
     it('add multiple keys in parallel', async () => {
         const assignments = range(10).map(() => {
-            return { key: GroupKey.generate(), publisherId: randomEthereumAddress() }
+            return { key: GroupKey.generate(), publisherId: randomUserId() }
         })
         await Promise.all(assignments.map(({ key, publisherId }) => store.set(key.id, publisherId, key.data)))
         for (const assignment of assignments) {
@@ -82,35 +70,12 @@ describe('LocalGroupKeyStore', () => {
         }
     })
 
-    /**
-     * Legacy keys refer to group keys migrated from a previous version of the client where group keys were not tied
-     * to a specific publisherId, therefore any publisherId for a given legacy key id is considered a match.
-     */
-    it('supports "legacy" keys', async () => {
-        const groupKey = GroupKey.generate()
-        const internalPersistence = await persistenceManager.getPersistence('EncryptionKeys')
-        await internalPersistence.set(`LEGACY::${groupKey.id}`, Buffer.from(groupKey.data).toString('hex'))
-        expect(await store.get(groupKey.id, randomEthereumAddress())).toEqual(groupKey)
-    })
-
-    it('"normal" keys have precedence over "legacy" keys', async () => {
-        const keyId = GroupKey.generate().id
-        const legacyKey = new GroupKey(keyId, crypto.randomBytes(32))
-        const normalKey = new GroupKey(keyId, crypto.randomBytes(32))
-        const internalPersistence = await persistenceManager.getPersistence('EncryptionKeys')
-        await internalPersistence.set(`LEGACY::${legacyKey.id}`, Buffer.from(legacyKey.data).toString('hex'))
-        await internalPersistence.set(`${publisherId}::${normalKey.id}`, Buffer.from(normalKey.data).toString('hex'))
-
-        expect(await store.get(keyId, publisherId)).toEqual(normalKey)
-        expect(await store.get(keyId, randomEthereumAddress())).toEqual(legacyKey)
-    })
-
     describe('latest encryptionKey id', () => {
         const streamId = toStreamID('/foobar', randomEthereumAddress())
         it('add and get key', async () => {
             await store.setLatestEncryptionKeyId('keyId', publisherId, streamId)
             expect(await store.getLatestEncryptionKeyId(publisherId, streamId)).toEqual('keyId')
-            expect(await store.getLatestEncryptionKeyId(randomEthereumAddress(), streamId)).toBeUndefined()
+            expect(await store.getLatestEncryptionKeyId(randomUserId(), streamId)).toBeUndefined()
             expect(await store.getLatestEncryptionKeyId(publisherId, toStreamID('foobar'))).toBeUndefined()
         })
     })