+++ packages/test-utils/src/customMatcherTypes.ts
@@ -1,0 +1,11 @@
+import { CustomMatchers } from './customMatchers'
+
+// we could ES2015 module syntax (https://jestjs.io/docs/expect#expectextendmatchers),
+// but the IDE doesn't find custom matchers if we do that
+declare global {
+    // eslint-disable-next-line @typescript-eslint/no-namespace
+    namespace jest {
+        interface Expect extends CustomMatchers {}
+        interface Matchers<R> extends CustomMatchers<R> {}
+    }
+}