+++ packages/trackerless-network/test/end-to-end/proxy-and-full-node.test.ts
@@ -1,4 +1,4 @@
-import { randomEthereumAddress } from '@streamr/test-utils'
+import { randomEthereumAddress, randomUserId } from '@streamr/test-utils'
 import { StreamPartID, StreamPartIDUtils, hexToBinary, utf8ToBinary, waitForEvent3 } from '@streamr/utils'
 import { NetworkNode, createNetworkNode } from '../../src/NetworkNode'
 import {
@@ -10,10 +10,10 @@ import {
 } from '../../src/proto/packages/trackerless-network/protos/NetworkRpc'
 import { createMockPeerDescriptor } from '../utils/utils'
 
-const PROXIED_NODE_USER_ID = randomEthereumAddress()
+const PROXIED_NODE_USER_ID = randomUserId()
 
 const createMessage = (streamPartId: StreamPartID): StreamMessage => {
-    return { 
+    return {
         messageId: {
             streamId: StreamPartIDUtils.getStreamID(streamPartId),
             streamPartition: StreamPartIDUtils.getStreamPartition(streamPartId),