+++ packages/trackerless-network/test/integration/ContentDeliveryManager.test.ts
@@ -3,11 +3,11 @@ import {
     Simulator,
     SimulatorTransport
 } from '@streamr/dht'
-import { randomEthereumAddress } from '@streamr/test-utils'
-import { StreamPartIDUtils, wait, waitForCondition, waitForEvent3 } from '@streamr/utils'
+import { StreamPartIDUtils, waitForCondition, waitForEvent3, wait } from '@streamr/utils'
 import { ContentDeliveryManager, Events } from '../../src/logic/ContentDeliveryManager'
 import { ControlLayerNode } from '../../src/logic/ControlLayerNode'
 import { createMockPeerDescriptor, createStreamMessage } from '../utils/utils'
+import { randomUserId } from '@streamr/test-utils'
 
 describe('ContentDeliveryManager', () => {
 
@@ -25,18 +25,12 @@ describe('ContentDeliveryManager', () => {
     const msg = createStreamMessage(
         JSON.stringify({ hello: 'WORLD' }),
         STREAM_PART_ID,
-        randomEthereumAddress()
+        randomUserId()
     )
-
-    afterEach(async () => {
-        await Promise.all([
-            manager1.destroy(),
-            manager2.destroy()
-        ])
-    })
+    let simulator: Simulator
 
     beforeEach(async () => {
-        const simulator = new Simulator()
+        simulator = new Simulator()
         transport1 = new SimulatorTransport(peerDescriptor1, simulator)
         await transport1.start()
         transport2 = new SimulatorTransport(peerDescriptor2, simulator)
@@ -68,6 +62,18 @@ describe('ContentDeliveryManager', () => {
         manager1.setStreamPartEntryPoints(STREAM_PART_ID, [peerDescriptor1])
         await manager2.start(controlLayerNode2, transport2, transport2)
         manager2.setStreamPartEntryPoints(STREAM_PART_ID, [peerDescriptor1])
+    })
+
+    afterEach(async () => {
+        await Promise.all([
+            manager1.destroy(),
+            manager2.destroy(),
+            controlLayerNode1.stop(),
+            controlLayerNode2.stop(),
+            transport1.stop(),
+            transport2.stop()
+        ])
+        simulator.stop()
     })
 
     it('starts', async () => {
@@ -112,7 +118,7 @@ describe('ContentDeliveryManager', () => {
         const msg2 = createStreamMessage(
             JSON.stringify({ hello: 'WORLD' }),
             streamPartId2,
-            randomEthereumAddress()
+            randomUserId()
         )
         await Promise.all([
             waitForEvent3<Events>(manager1, 'newMessage'),