+++ packages/trackerless-network/test/unit/ContentDeliveryLayerNode.test.ts
@@ -105,6 +105,8 @@ describe('ContentDeliveryLayerNode', () => {
         nodeWithRtt.setRtt(100)
         const info = contentDeliveryLayerNode.getInfo()
         expect(info[0].rtt).toEqual(100)
+        expect(info[0].peerDescriptor).toEqual(nodeWithRtt.getPeerDescriptor())
         expect(info[1].rtt).toBeUndefined()
+        expect(info[1].peerDescriptor).toEqual(nodeWithoutRtt.getPeerDescriptor())
     })
 })